requirements.txt
# CryptoDistro 2.0 — Dependencies

# Binance official SDK
binance-connector>=3.0.0

# Telegram bot
python-telegram-bot>=21.0

# HTTP requests (Noones API, Boltz API, Binance P2P)
httpx>=0.27.0

# Config
pyyaml>=6.0

# Intelligence agent (Cerebras via OpenAI-compatible SDK)
openai>=1.30.0

# LND gRPC (Phase 2 — install when needed)
# grpcio>=1.60.0
# protobuf>=4.25.0
